+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## AnyMind Angular API library changelog
+<a name="Jan 21, 2019"></a>
+### Jan 21, 2019
+* new generator, swagger generator is not used any more, custom scripts. 
+
 <a name="Sep 19, 2018"></a>
 ### Sep 19, 2018
 * Update packager version